refactor selectPage to select
