+++ lib/queries/crud.js
@@ -1,7 +1,7 @@
 import removeOneQuery from '../queries/removeOne';
 import insertOneQuery from '../queries/insertOne';
 import selectOneQuery from '../queries/selectOne';
-import selectPageQuery from '../queries/selectPage';
+import selectQuery from '../queries/select';
 import updateOneQuery from '../queries/updateOne';
 import batchInsertQuery from '../queries/batchInsert';
 import batchRemoveQuery from '../queries/batchRemove';
@@ -10,7 +10,7 @@ import countAllQuery from '../queries/countAll';
 export default function ({
     table,
     fields,
-    idFields = ['id'],
+    primaryKey = 'id',
     returnFields = fields,
     updatableFields = fields,
     searchableFields,
@@ -19,20 +19,20 @@ export default function ({
     withQuery,
     allowPrimaryKeyUpdate,
 }) {
-    const removeOne = removeOneQuery({ table, primaryKey: idFields });
+    const removeOne = removeOneQuery({ table, primaryKey });
     const insertOne = insertOneQuery({ table, fields, returnFields });
-    const selectOne = selectOneQuery({ table, idFields, returnFields });
-    const selectPage = selectPageQuery({ table, idFields, returnFields, searchableFields, specificSorts, groupByFields, withQuery });
-    const updateOne = updateOneQuery({ table, updatableFields, primaryKey: idFields, returnFields, allowPrimaryKeyUpdate });
+    const selectOne = selectOneQuery({ table, primaryKey, returnFields });
+    const select = selectQuery({ table, primaryKey, returnFields, searchableFields, specificSorts, groupByFields, withQuery });
+    const updateOne = updateOneQuery({ table, updatableFields, primaryKey, returnFields, allowPrimaryKeyUpdate });
     const batchInsert = batchInsertQuery({ table, fields, returnFields });
-    const batchRemove = batchRemoveQuery({ table, returnFields, primaryKey: idFields });
+    const batchRemove = batchRemoveQuery({ table, returnFields, primaryKey });
     const countAll = countAllQuery({ table });
 
     const queries = {
         removeOne,
         insertOne,
         selectOne,
-        selectPage,
+        select,
         updateOne,
         batchInsert,
         batchRemove,
