+++ lib/queries/crud.spec.js
@@ -18,7 +18,7 @@ describe('crud', () => {
             'removeOne',
             'insertOne',
             'selectOne',
-            'selectPage',
+            'select',
             'updateOne',
             'batchInsert',
             'batchRemove',
