+++ lib/queries/index.js
@@ -3,7 +3,7 @@ export { default as removeOneQuery } from './removeOne';
 export { default as remove } from './remove';
 export { default as insertOneQuery } from './insertOne';
 export { default as selectOneQuery } from './selectOne';
-export { default as selectPageQuery } from './selectPage';
+export { default as selectQuery } from './select';
 export { default as updateOneQuery } from './updateOne';
 export { default as updateQuery } from './update';
 export { default as batchInsertQuery } from './batchInsert';
