+++ lib/queries/select.js
@@ -4,14 +4,14 @@ import flattenParameters from './flattenParameters';
 
 export default ({
     table,
-    idFields: ids = ['id'],
+    primaryKey: ids = 'id',
     returnFields,
     searchableFields = returnFields || [],
     specificSorts = {},
     groupByFields = [],
     withQuery = table.indexOf('JOIN') !== -1,
 }) => {
-    const idFields = [].concat(ids);
+    const primaryKey = [].concat(ids);
     const select = returnFields.length ? returnFields.join(', ') : '*';
 
     return ({ limit, offset, filters = {}, sort, sortDir } = { filters: {} }) => {
@@ -33,7 +33,7 @@ ${sql.trim()}
         const parameters = flattenParameters(sanitizeParameter(searchableFields, filters));
 
         // always sort by ids to avoid randomness in case of identical sortField value
-        const sortQuery = [idFields.map(idField => `${idField} ASC`).join(', ')];
+        const sortQuery = [primaryKey.map(idField => `${idField} ASC`).join(', ')];
         if (sort) {
             const normalizedSort = sort.toLowerCase();
             if (specificSorts && specificSorts.hasOwnProperty(normalizedSort)) {
@@ -60,6 +60,6 @@ ${sql.trim()}
             parameters.offset = offset || 0;
         }
 
-        return { sql, parameters };
+        return { sql, parameters, returnOne: limit === 1 };
     };
 };
