+++ lib/queries/select.spec.js
@@ -1,12 +1,12 @@
 import { assert } from 'chai';
 
-import selectPage from './selectPage';
+import select from './select';
 
-describe('selectPage', () => {
+describe('select', () => {
     it('should use a simple query if querying on a single table', () => {
-        const { sql, parameters } = selectPage({
+        const { sql, parameters } = select({
             table: 'table',
-            idFields: ['id'],
+            primaryKey: ['id'],
             returnFields: ['field1', 'field2'],
         })();
 
@@ -15,9 +15,9 @@ describe('selectPage', () => {
     });
 
     it('should add a group by clause if at least on groupByFields is given', () => {
-        const { sql, parameters } = selectPage({
+        const { sql, parameters } = select({
             table: 'table',
-            idFields: ['id'],
+            primaryKey: ['id'],
             returnFields: ['field1', 'field2'],
             groupByFields: ['field1', 'field2'],
         })();
@@ -27,9 +27,9 @@ describe('selectPage', () => {
     });
 
     it('should use a "WITH result AS" query if querying on a joined table', () => {
-        const { sql, parameters } = selectPage({
+        const { sql, parameters } = select({
             table: 'table1 JOIN table2 ON table1.table2_id table2.id',
-            idFields: ['id'],
+            primaryKey: ['id'],
             returnFields: ['field1', 'field2'],
         })();
 
@@ -42,9 +42,9 @@ SELECT field1, field2 FROM table1 JOIN table2 ON table1.table2_id table2.id
     });
 
     it('should use a "WITH result AS" query if enabling the withQuery extraOptions', () => {
-        const { sql, parameters } = selectPage({
+        const { sql, parameters } = select({
             table: 'table1 JOIN table2 ON table1.table2_id table2.id',
-            idFields: ['id'],
+            primaryKey: ['id'],
             returnFields: ['field1', 'field2'],
             withQuery: true,
         })();
@@ -58,9 +58,9 @@ SELECT field1, field2 FROM table1 JOIN table2 ON table1.table2_id table2.id
     });
 
     it('should use filters in WHERE query', () => {
-        const { sql, parameters } = selectPage({
+        const { sql, parameters } = select({
             table: 'table',
-            idFields: ['id'],
+            primaryKey: ['id'],
             returnFields: ['field1', 'field2'],
         })({ filters: { field1: 'value', field2: 'other value' } });
 
@@ -72,9 +72,9 @@ SELECT field1, field2 FROM table1 JOIN table2 ON table1.table2_id table2.id
     });
 
     it('should use IN query when receiving an array of value as a filter parameter', () => {
-        const { sql, parameters } = selectPage({
+        const { sql, parameters } = select({
             table: 'table',
-            idFields: ['id'],
+            primaryKey: ['id'],
             returnFields: ['field1', 'field2'],
         })({ filters: { field1: ['value', 'other value'] } });
 
@@ -86,9 +86,9 @@ SELECT field1, field2 FROM table1 JOIN table2 ON table1.table2_id table2.id
     });
 
     it('should limit and skip if given corresponding parameter', () => {
-        const { sql, parameters } = selectPage({
+        const { sql, parameters } = select({
             table: 'table',
-            idFields: ['id'],
+            primaryKey: ['id'],
             returnFields: ['field1', 'field2'],
         })({ limit: 50, offset: 100 });
 
@@ -97,9 +97,9 @@ SELECT field1, field2 FROM table1 JOIN table2 ON table1.table2_id table2.id
     });
 
     it('should sort using sort and sortDir', () => {
-        const { sql, parameters } = selectPage({
+        const { sql, parameters } = select({
             table: 'table',
-            idFields: ['id'],
+            primaryKey: ['id'],
             returnFields: ['field1', 'field2'],
         })({ sort: 'field2', sortDir: 'DESC' });
 
@@ -108,9 +108,9 @@ SELECT field1, field2 FROM table1 JOIN table2 ON table1.table2_id table2.id
     });
 
     it('should use specificsort if it is set and used', () => {
-        const { sql, parameters } = selectPage({
+        const { sql, parameters } = select({
             table: 'table',
-            idFields: ['id'],
+            primaryKey: ['id'],
             returnFields: ['field1', 'field2'],
             specificSorts: { level: ['master', 'expert', 'novice'] },
         })({ sort: 'level', sortDir: 'DESC' });
